+++ event-sourcing/event-publisher/event-publisher-timer/src/test/java/uk/gov/justice/services/eventsourcing/eventpublishing/LinkedEventPublisherTest.java
@@ -21,6 +21,7 @@ import uk.gov.justice.services.eventsourcing.publishedevent.jdbc.CompatibilityMo
 import uk.gov.justice.services.eventsourcing.publishedevent.jdbc.EventPublishingRepository;
 import uk.gov.justice.services.eventsourcing.publisher.jms.EventPublisher;
 import uk.gov.justice.services.eventsourcing.repository.jdbc.event.LinkedEvent;
+import uk.gov.justice.services.eventsourcing.repository.jdbc.event.MissingEventNumberException;
 import uk.gov.justice.services.messaging.JsonEnvelope;
 
 import java.util.UUID;
@@ -56,6 +57,7 @@ public class LinkedEventPublisherTest {
     @Test
     public void shouldGetEventIdFromPublishQueueFindTheEventInEventLogAddPreviousAndNextEventNumberToMetadataAndPublish() throws Exception {
         final UUID eventId = randomUUID();
+        final Long eventNumber = 23L;
 
         final LinkedEvent linkedEvent = mock(LinkedEvent.class);
         final JsonEnvelope linkedJsonEnvelope = mock(JsonEnvelope.class);
@@ -64,6 +66,7 @@ public class LinkedEventPublisherTest {
         when(eventPublishingRepository.findEventFromEventLog(eventId)).thenReturn(of(linkedEvent));
         when(linkedJsonEnvelopeCreator.createLinkedJsonEnvelopeFrom(linkedEvent)).thenReturn(linkedJsonEnvelope);
         when(eventPublishingWorkerConfig.shouldAlsoInsertEventIntoPublishedEventTable()).thenReturn(true);
+        when(linkedEvent.getEventNumber()).thenReturn(of(eventNumber));
 
         assertThat(linkedEventPublisher.publishNextNewEvent(), is(true));
 
@@ -75,6 +78,7 @@ public class LinkedEventPublisherTest {
         inOrder.verify(eventPublisher).publish(linkedJsonEnvelope);
         inOrder.verify(eventPublishingRepository).setIsPublishedFlag(eventId, true);
         inOrder.verify(compatibilityModePublishedEventRepository).insertIntoPublishedEvent(linkedJsonEnvelope);
+        inOrder.verify(compatibilityModePublishedEventRepository).setEventNumberSequenceTo(eventNumber);
     }
 
     @Test
@@ -99,6 +103,27 @@ public class LinkedEventPublisherTest {
         inOrder.verify(eventPublisher).publish(linkedJsonEnvelope);
         inOrder.verify(eventPublishingRepository).setIsPublishedFlag(eventId, true);
         inOrder.verify(compatibilityModePublishedEventRepository, never()).insertIntoPublishedEvent(linkedJsonEnvelope);
+    }
+
+    @Test
+    public void shouldThrowMissingEventNumberExceptionIfEventInEvenLongTableHasNullEventNumber() throws Exception {
+        final UUID eventId = fromString("15892431-6cb7-43d4-9d55-1aef19a6d087");
+
+        final LinkedEvent linkedEvent = mock(LinkedEvent.class);
+        final JsonEnvelope linkedJsonEnvelope = mock(JsonEnvelope.class);
+
+        when(eventPublishingRepository.popNextEventIdFromPublishQueue()).thenReturn(of(eventId));
+        when(eventPublishingRepository.findEventFromEventLog(eventId)).thenReturn(of(linkedEvent));
+        when(linkedJsonEnvelopeCreator.createLinkedJsonEnvelopeFrom(linkedEvent)).thenReturn(linkedJsonEnvelope);
+        when(eventPublishingWorkerConfig.shouldAlsoInsertEventIntoPublishedEventTable()).thenReturn(true);
+        when(linkedEvent.getEventNumber()).thenReturn(empty());
+
+        final MissingEventNumberException missingEventNumberException = assertThrows(
+                MissingEventNumberException.class,
+                () -> linkedEventPublisher.publishNextNewEvent());
+
+        assertThat(missingEventNumberException.getMessage(), is("Event with id '15892431-6cb7-43d4-9d55-1aef19a6d087' has null event_number in event_log table"));
+
     }
 
     @Test